+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/WatchConnectionManager.java
@@ -33,6 +33,7 @@ import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 
@@ -45,10 +46,12 @@ import static java.net.HttpURLConnection.HTTP_UNAVAILABLE;
 public class WatchConnectionManager<T extends HasMetadata, L extends KubernetesResourceList<T>>
     extends AbstractWatchManager<T> {
 
+  public static final int BACKOFF_MAX_EXPONENT = 5;
+
   private static final Logger logger = LoggerFactory.getLogger(WatchConnectionManager.class);
 
   protected WatcherWebSocketListener<T> listener;
-  private CompletableFuture<WebSocket> websocketFuture;
+  private volatile CompletableFuture<WebSocket> websocketFuture;
   private WebSocket websocket;
 
   private volatile boolean ready;
@@ -78,20 +81,21 @@ public class WatchConnectionManager<T extends HasMetadata, L extends KubernetesR
       final ListOptions listOptions, final Watcher<T> watcher, final int reconnectInterval, final int reconnectLimit,
       long websocketTimeout) throws MalformedURLException {
     // Default max 32x slowdown from base interval
-    this(client, baseOperation, listOptions, watcher, reconnectInterval, reconnectLimit, websocketTimeout, 5);
+    this(client, baseOperation, listOptions, watcher, reconnectInterval, reconnectLimit, websocketTimeout,
+        BACKOFF_MAX_EXPONENT);
   }
 
   @Override
   protected synchronized void closeRequest() {
     closeWebSocket(websocket);
-    if (this.websocketFuture != null) {
-      this.websocketFuture.whenComplete((w, t) -> {
+    Optional.ofNullable(this.websocketFuture).ifPresent(theFuture -> {
+      this.websocketFuture = null;
+      theFuture.whenComplete((w, t) -> {
         if (w != null) {
           closeWebSocket(w);
         }
       });
-      websocketFuture = null;
-    }
+    });
   }
 
   synchronized WatcherWebSocketListener<T> getListener() {
@@ -100,6 +104,14 @@ public class WatchConnectionManager<T extends HasMetadata, L extends KubernetesR
 
   public CompletableFuture<WebSocket> getWebsocketFuture() {
     return websocketFuture;
+  }
+
+  @Override
+  protected void onMessage(String message) {
+    // for consistency we only want to process the message when we're open
+    if (this.websocketFuture != null) {
+      super.onMessage(message);
+    }
   }
 
   @Override