+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/WatcherWebSocketListener.java
@@ -57,7 +57,6 @@ class WatcherWebSocketListener<T extends HasMetadata> implements WebSocket.Liste
     } finally {
       webSocket.request();
     }
-    webSocket.request();
   }
 
   @Override